requirements.txt
asgiref==3.6.0
astroid==2.15.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.2
boto==2.49.0
boto3==1.26.107
botocore==1.29.107
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==40.0.2
dill==0.3.6
Django==4.2
django-cors-headers==3.14.0
django-filter==22.1
django-socket-server==0.0.4
django-storages==1.13.2
djangorestframework==3.14.0
docutils==0.19
hyperlink==21.0.0
idna==3.4
image==1.5.33
incremental==22.10.0
isort==5.12.0
jmespath==1.0.1
lazy-object-proxy==1.9.0
Markdown==3.4.1
mccabe==0.7.0
Pillow==9.5.0
platformdirs==3.2.0
psycopg==3.1.8
pur==7.1.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==1.7.1
pylint==2.17.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomlkit==0.11.7
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
wrapt==1.15.0
zope.interface==6.0
